refactor(app): extract message ingestion
